backend/services/authenticator/tests/requirements.txt
pytest
pyjwt[crypto]
argon2-cffi
fastapi
httpx
psycopg[binary]
betterproto[compiler]
grpclib